third_party/WebKit/LayoutTests/webaudio/internals/mediaelementaudiosourcenode-gc-expected.txt

CONSOLE WARNING: line 22: The Web Audio autoplay policy will be re-enabled in Chrome 70 (October 2018). Please check that your website is compatible with it. https://goo.gl/7K7WLu
Tests garbage collection of MediaElementAudioSourceNode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS MediaElementAudioSourceNode survived garbage collection.
PASS successfullyParsed is true

TEST COMPLETE
